+++ src/test/java/com/puppycrawl/tools/checkstyle/checks/indentation/CommentsIndentationCheckTest.java
@@ -133,7 +133,7 @@ public class CommentsIndentationCheckTest extends AbstractModuleTestSupport {
     }
 
     @Test
-    public void testCommentIsInsideSwitchBlock() throws Exception {
+    public void testCommentIsInsideSwitchBlockOne() throws Exception {
         final String[] expected = {
             "27:13: " + getCheckMessage(MSG_KEY_BLOCK, 28, 12, 16),
             "33:20: " + getCheckMessage(MSG_KEY_SINGLE, "32, 34", 19, "16, 12"),
@@ -143,20 +143,47 @@ public class CommentsIndentationCheckTest extends AbstractModuleTestSupport {
             "67:23: " + getCheckMessage(MSG_KEY_SINGLE, 66, 22, 16),
             "76:15: " + getCheckMessage(MSG_KEY_SINGLE, "73, 77", 14, "12, 16"),
             "96:25: " + getCheckMessage(MSG_KEY_SINGLE, 97, 24, 20),
-            "121:16: " + getCheckMessage(MSG_KEY_SINGLE, "120, 122", 15, "17, 12"),
-            "133:9: " + getCheckMessage(MSG_KEY_SINGLE, 134, 8, 12),
-            "146:5: " + getCheckMessage(MSG_KEY_SINGLE, 147, 4, 8),
-            "165:19: " + getCheckMessage(MSG_KEY_SINGLE, "164, 166", 18, "16, 12"),
-            "208:5: " + getCheckMessage(MSG_KEY_SINGLE, "207, 209", 4, "12, 12"),
-            "211:23: " + getCheckMessage(MSG_KEY_SINGLE, "210, 214", 22, "16, 12"),
-            "212:21: " + getCheckMessage(MSG_KEY_SINGLE, "210, 214", 20, "16, 12"),
-            "213:18: " + getCheckMessage(MSG_KEY_SINGLE, "210, 214", 17, "16, 12"),
-            "237:7: " + getCheckMessage(MSG_KEY_SINGLE, "236, 238", 6, "12, 12"),
-            "284:12: " + getCheckMessage(MSG_KEY_BLOCK, "283, 287", 11, "16, 12"),
-            "289:12: " + getCheckMessage(MSG_KEY_SINGLE, "288, 290", 11, "16, 12"),
-            "319:1: " + getCheckMessage(MSG_KEY_SINGLE, "320", 0, "8"),
         };
-        final String testInputFile = "InputCommentsIndentationInSwitchBlock.java";
+        final String testInputFile = "InputCommentsIndentationInSwitchBlockOne.java";
+        verifyWithInlineConfigParser(getPath(testInputFile), expected);
+    }
+
+    @Test
+    public void testCommentIsInsideSwitchBlockTwo() throws Exception {
+        final String[] expected = {
+            "18:25: " + getCheckMessage(MSG_KEY_SINGLE, 19, 24, 20),
+            "43:16: " + getCheckMessage(MSG_KEY_SINGLE, "42, 44", 15, "17, 12"),
+            "55:9: " + getCheckMessage(MSG_KEY_SINGLE, 56, 8, 12),
+            "68:5: " + getCheckMessage(MSG_KEY_SINGLE, 69, 4, 8),
+            "87:19: " + getCheckMessage(MSG_KEY_SINGLE, "86, 88", 18, "16, 12"),
+        };
+        final String testInputFile = "InputCommentsIndentationInSwitchBlockTwo.java";
+        verifyWithInlineConfigParser(getPath(testInputFile), expected);
+    }
+
+    @Test
+    public void testCommentIsInsideSwitchBlockThree() throws Exception {
+        final String[] expected = {
+            "18:25: " + getCheckMessage(MSG_KEY_SINGLE, 19, 24, 20),
+            "45:5: " + getCheckMessage(MSG_KEY_SINGLE, "44, 46", 4, "12, 12"),
+            "48:23: " + getCheckMessage(MSG_KEY_SINGLE, "47, 51", 22, "16, 12"),
+            "49:21: " + getCheckMessage(MSG_KEY_SINGLE, "47, 51", 20, "16, 12"),
+            "50:18: " + getCheckMessage(MSG_KEY_SINGLE, "47, 51", 17, "16, 12"),
+            "74:7: " + getCheckMessage(MSG_KEY_SINGLE, "73, 75", 6, "12, 12"),
+        };
+        final String testInputFile = "InputCommentsIndentationInSwitchBlockThree.java";
+        verifyWithInlineConfigParser(getPath(testInputFile), expected);
+    }
+
+    @Test
+    public void testCommentIsInsideSwitchBlockFour() throws Exception {
+        final String[] expected = {
+            "18:25: " + getCheckMessage(MSG_KEY_SINGLE, 19, 24, 20),
+            "34:12: " + getCheckMessage(MSG_KEY_BLOCK, "33, 37", 11, "16, 12"),
+            "39:12: " + getCheckMessage(MSG_KEY_SINGLE, "38, 40", 11, "16, 12"),
+            "69:1: " + getCheckMessage(MSG_KEY_SINGLE, "70", 0, "8"),
+        };
+        final String testInputFile = "InputCommentsIndentationInSwitchBlockFour.java";
         verifyWithInlineConfigParser(getPath(testInputFile), expected);
     }
 
@@ -277,7 +304,7 @@ public class CommentsIndentationCheckTest extends AbstractModuleTestSupport {
     }
 
     @Test
-    public void testMultiblockStructures() throws Exception {
+    public void testMultiblockStructuresOne() throws Exception {
         final String[] expected = {
             "19:9: " + getCheckMessage(MSG_KEY_SINGLE, 18, 8, 12),
             "25:17: " + getCheckMessage(MSG_KEY_SINGLE, "24, 26", 16, "12, 8"),
@@ -294,11 +321,19 @@ public class CommentsIndentationCheckTest extends AbstractModuleTestSupport {
             "103:9: " + getCheckMessage(MSG_KEY_SINGLE, 102, 8, 12),
             "109:1: " + getCheckMessage(MSG_KEY_SINGLE, "108, 110", 0, "12, 8"),
             "114:17: " + getCheckMessage(MSG_KEY_SINGLE, "113, 115", 16, "12, 8"),
-            "124:9: " + getCheckMessage(MSG_KEY_SINGLE, 123, 8, 12),
-            "130:17: " + getCheckMessage(MSG_KEY_SINGLE, "129, 131", 16, "12, 8"),
-            "135:1: " + getCheckMessage(MSG_KEY_SINGLE, "134, 136", 0, "12, 8"),
         };
-        final String testInputFile = "InputCommentsIndentationInMultiblockStructures.java";
+        final String testInputFile = "InputCommentsIndentationInMultiblockStructuresOne.java";
+        verifyWithInlineConfigParser(getPath(testInputFile), expected);
+    }
+
+    @Test
+    public void testMultiblockStructuresTwo() throws Exception {
+        final String[] expected = {
+            "20:9: " + getCheckMessage(MSG_KEY_SINGLE, 19, 8, 12),
+            "26:17: " + getCheckMessage(MSG_KEY_SINGLE, "25, 27", 16, "12, 8"),
+            "31:1: " + getCheckMessage(MSG_KEY_SINGLE, "30, 32", 0, "12, 8"),
+        };
+        final String testInputFile = "InputCommentsIndentationInMultiblockStructuresTwo.java";
         verifyWithInlineConfigParser(getPath(testInputFile), expected);
     }
