+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/indentation/commentsindentation/InputCommentsIndentationInSwitchBlockThree.java
@@ -1,0 +1,113 @@
+/*
+CommentsIndentation
+tokens = (default)SINGLE_LINE_COMMENT, BLOCK_COMMENT_BEGIN
+
+
+*/
+
+package com.puppycrawl.tools.checkstyle.checks.indentation.commentsindentation;
+
+public class InputCommentsIndentationInSwitchBlockThree {
+
+    private static void foo1() {
+        if (true) {
+            switch(1) {
+                case 0:
+
+                case 1:
+                        // violation '.* incorrect .* level 24, expected is 20,.*same.* as line 19.'
+                    int b = 10;
+                default:
+                 // comment
+            }
+
+        }
+    }
+
+    public void foo7() {
+        int a = 2;
+        String s = "";
+        switch (a) {
+            // comment
+            // comment
+            // comment
+            case 1:
+            case 2:
+                // comment
+                // comment
+                foo1();
+                // comment
+            case 3:
+                // comment
+                // comment
+                // comment
+            case 4:
+    // violation '.* incorrect .* level 4, expected is 12, 12, .* same .* as line 44, 46.'
+            case 5:
+                s.toString().toString().toString();
+                      // violation '.*incorrect.* level 22, expected is 16, 12,.*as line 47, 51.'
+                    // violation '.* incorrect .* level 20, expected is 16, 12,.*as line 47, 51.'
+                 // violation '.* incorrect .* level 17, expected is 16, 12, .* as line 47, 51.'
+            default:
+        }
+    }
+
+    public void foo8() {
+        int a = 2;
+        String s = "";
+        switch (a) {
+            // comment
+            // comment
+            // comment
+            case 1:
+            case 2:
+                // comment
+                // comment
+                foo1();
+                // comment
+            case 3:
+                // comment
+                // comment
+                s.toString().toString().toString();
+                // comment
+            case 4:
+      // violation '.* incorrect .* level 6, expected is 12, 12, .* same .* as line 73, 75.'
+            default:
+        }
+    }
+
+    public void foo9() {
+        int a = 5;
+        switch (a) {
+            case 1:
+            case 2:
+                // comment
+        }
+    }
+
+    public void foo10() {
+        int a = 5;
+        switch (a) {
+            case 1:
+            default:
+                // comment
+        }
+    }
+
+    public void foo11() {
+        int a = 5;
+        switch (a) {
+            case 1:
+            case 2:
+                // comment
+        }
+    }
+
+    public void foo12() {
+        int a = 5;
+        switch (a) {
+            // comment
+            case 1:
+        }
+    }
+}
